refactor: move uint128 to own file
